requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.4
schedule==1.1.0
py==1.10.0
pytest==3.0.2
wheel==0.24.0
python-telegram-bot==13.5
